Созданы классы для делегации: 1. Создан класс ApplicationManager, в который делегируется управление приложением из класса TestBase. 2. Созданы классы-помощники GroupHelper, NavigationHelper и SessionHelper, в которые делегируется логика соответствующих действий: управление группой, навигация, вход в систему/сессия. 3. Для упорядочивания назначения классов созданы дополнительные пакеты: appManager для управления приложением, tests для хранения тестовых процедур, model для описание предметной области.
