+++ src/test/java/ru/rt/client/model/GroupData.java
@@ -1,4 +1,4 @@
-package ru.rt.client;
+package ru.rt.client.model;
 
 public class GroupData {
     private final String name;
