+++ src/test/java/ru/rt/client/tests/GroupRemovalTest.java
@@ -1,0 +1,14 @@
+package ru.rt.client.tests;
+
+import org.testng.annotations.Test;
+
+public class GroupRemovalTest extends TestBase {
+
+    @Test
+    public void testGroupRemoval() throws Exception {
+        app.getNavigationHelper().navToGroupPage();
+        app.getGroupHelper().selectGroup();
+        app.getGroupHelper().deleteSelectedGroup();
+    }
+
+}
